+++ WGLMakie/src/lines.jl
@@ -1,10 +1,10 @@
 function serialize_three(scene::Scene, plot::Union{Lines, LineSegments})
-    Makie.@converted_attribute plot (linewidth, color)
+    Makie.@converted_attribute plot (linewidth,)
     uniforms = Dict(
-        :pattern_length => 1f0,
         :model => plot.model,
-        :is_valid => Vec4f(1),
     )
+
+    color = plot.calculated_colors
     attributes = Dict{Symbol, Any}(
         :linepoint => lift(serialize_buffer_attribute, plot[1])
     )
@@ -20,10 +20,8 @@ function serialize_three(scene::Scene, plot::Union{Lines, LineSegments})
         :name => string(Makie.plotkey(plot)) * "-" * string(objectid(plot)),
         :visible => plot.visible,
         :uuid => js_uuid(plot),
-        :plot_type => :lines,
+        :plot_type => plot isa LineSegments ? "linesegments" : "lines",
         :cam_space => plot.space[],
-        :is_linesegments => plot isa LineSegments,
-
         :uniforms => serialize_uniforms(uniforms),
         :attributes => attributes
     )
